Add clang-toolchain image
This is a very simple image that supplies clang-format, git, and
parallel to facilitate running clang-format checks in CI. This also lays
the foundation for running clang-tidy if we need to in the future.

Also added is configuration to test, build, and push in Concourse.

Steps to locally run tests are documented in README.